Level Design / Fix
- Dragon 원거리 공격 오류 fix
- 1~7 난이도 하향 조절

+++ Assets/Scripts/#3LobbyScene/Quest/RepeatQuest.cs
@@ -20,7 +20,7 @@ public class RepeatQuest : MonoBehaviour
         clearAmount = BackEndGameData.Instance.UserQuestData.repeatQuest[questNum];
 
         clearAmountText.text = clearAmount.ToString() + "/" + targetAmount.ToString();
-        progressImage.fillAmount = clearAmount / targetAmount;
+        progressImage.fillAmount = (float)clearAmount / (float)targetAmount;
         QuestManager.instance.AlarmCheck();
         UpdateBtn();
     }
